drm/msm/sde: Skip heavy autorefresh checks when it's not enabled.
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: Iee862cff2d1ab904941a0bce3e55bc6deac94f0c
Signed-off-by: Anushek Prasal <anushekprasal@gmail.com>